bloom-extension/backend/requirements.txt
numpy==1.24.4
httpx<0.25.0,>=0.23.0
openai==0.28.1
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
chromadb==0.4.18
pymupdf==1.23.6
python-docx==0.8.11
python-dotenv==1.0.0
pydantic==2.5.0